refactor: rename modules to be private
